refactor: extract a separate test_hub.py file from test_spawn.py
